+++ sshd-core/src/main/java/org/apache/sshd/client/session/ClientConnectionService.java
@@ -23,12 +23,15 @@ import java.time.Duration;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
 import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicInteger;
 
 import org.apache.sshd.agent.common.AgentForwardSupport;
 import org.apache.sshd.common.FactoryManager;
 import org.apache.sshd.common.SshConstants;
 import org.apache.sshd.common.SshException;
+import org.apache.sshd.common.future.GlobalRequestFuture;
 import org.apache.sshd.common.io.IoWriteFuture;
+import org.apache.sshd.common.kex.KexState;
 import org.apache.sshd.common.session.Session;
 import org.apache.sshd.common.session.helpers.AbstractConnectionService;
 import org.apache.sshd.common.util.GenericUtils;
@@ -46,7 +49,10 @@ public class ClientConnectionService
         implements ClientSessionHolder {
     protected final String heartbeatRequest;
     protected final Duration heartbeatInterval;
-    protected final Duration heartbeatReplyMaxWait;
+    protected final int heartbeatMaxNoReply;
+
+    protected final AtomicInteger outstandingHeartbeats = new AtomicInteger();
+
     /** Non-null only if using the &quot;keep-alive&quot; request mechanism */
     protected ScheduledFuture<?> clientHeartbeat;
 
@@ -55,7 +61,44 @@ public class ClientConnectionService
 
         heartbeatRequest = CoreModuleProperties.HEARTBEAT_REQUEST.getRequired(this);
         heartbeatInterval = CoreModuleProperties.HEARTBEAT_INTERVAL.getRequired(this);
-        heartbeatReplyMaxWait = CoreModuleProperties.HEARTBEAT_REPLY_WAIT.getRequired(this);
+        heartbeatMaxNoReply = configureMaxNoReply();
+    }
+
+    protected int configureMaxNoReply() {
+        @SuppressWarnings("deprecation")
+        Duration timeout = CoreModuleProperties.HEARTBEAT_REPLY_WAIT.getOrNull(this);
+        if (timeout == null || GenericUtils.isNegativeOrNull(heartbeatInterval) || GenericUtils.isEmpty(heartbeatRequest)) {
+            return CoreModuleProperties.HEARTBEAT_NO_REPLY_MAX.getRequired(this).intValue();
+        }
+        // The deprecated timeout is configured explicitly. If the new no-reply-max is _not_ explicitly configured,
+        // set it from the timeout.
+        Integer noReplyValue = CoreModuleProperties.HEARTBEAT_NO_REPLY_MAX.getOrNull(this);
+        if (noReplyValue != null) {
+            return noReplyValue.intValue();
+        }
+        if (GenericUtils.isNegativeOrNull(timeout)) {
+            return 0;
+        }
+        if (timeout.compareTo(heartbeatInterval) >= 0) {
+            // Timeout is longer than the interval. With the previous system, that would have killed the session when
+            // the timeout was reached. A slow server that managed to return the reply just before the timeout expired
+            // would have delayed subsequent heartbeats. The new system will keep sending heartbeats with the given
+            // interval. Thus we can have timeout / interval heartbeats without reply if we want to approximate the
+            // timeout.
+            double timeoutSec = timeout.getSeconds() + (timeout.getNano() / 1_000_000_000.0);
+            double intervalSec = heartbeatInterval.getSeconds() + (heartbeatInterval.getNano() / 1_000_000_000.0);
+            double multiple = timeoutSec / intervalSec;
+            if (multiple >= Integer.MAX_VALUE - 1) {
+                return Integer.MAX_VALUE;
+            } else {
+                return (int) multiple + 1;
+            }
+        }
+        // Timeout is smaller than the interval. We want to have every heartbeat replied to.
+        return 1;
+        // This is an approximation. If no reply is forthcoming, the session will be killed after the interval. In the
+        // previous system, it would have been killed after the timeout. We _could_ code something to schedule a task
+        // that kills the session after the timeout and cancel that if we get a reply, but it seems a bit pointless.
     }
 
     @Override
@@ -82,6 +125,7 @@ public class ClientConnectionService
         if (!GenericUtils.isNegativeOrNull(heartbeatInterval) && GenericUtils.isNotEmpty(heartbeatRequest)) {
             stopHeartBeat();
 
+            outstandingHeartbeats.set(0);
             ClientSession session = getClientSession();
             FactoryManager manager = session.getFactoryManager();
             ScheduledExecutorService service = manager.getScheduledExecutorService();
@@ -117,26 +161,35 @@ public class ClientConnectionService
         }
 
         Session session = getSession();
+        if (session.getKexState() != KexState.DONE) {
+            // During KEX, global requests are delayed until after the key exchange is over. Don't count during KEX,
+            // otherwise a slow KEX might cause us to kill the session prematurely.
+            return false;
+        }
         try {
-            boolean withReply = !GenericUtils.isNegativeOrNull(heartbeatReplyMaxWait);
+            heartbeatCount.incrementAndGet();
+            boolean withReply = heartbeatMaxNoReply > 0;
+            int outstanding = outstandingHeartbeats.incrementAndGet();
+            if (withReply && heartbeatMaxNoReply < outstanding) {
+                throw new SshException("Got " + (outstanding - 1) + " heartbeat requests without reply");
+            }
             Buffer buf = session.createBuffer(
                     SshConstants.SSH_MSG_GLOBAL_REQUEST, heartbeatRequest.length() + Byte.SIZE);
             buf.putString(heartbeatRequest);
             buf.putBoolean(withReply);
 
+            // Even if we want a reply, we don't wait.
             if (withReply) {
-                Buffer reply = session.request(heartbeatRequest, buf, heartbeatReplyMaxWait);
-                if (reply != null) {
-                    if (log.isTraceEnabled()) {
-                        log.trace("sendHeartBeat({}) received reply size={} for request={}",
-                                session, reply.available(), heartbeatRequest);
-                    }
-                }
+                GlobalRequestFuture future = session.request(buf, heartbeatRequest, (cmd, buffer) -> {
+                    // We got something back. Don't care about success or failure. (In particular we may get here in
+                    // case the server responds SSH_MSG_UNIMPLEMENTED.)
+                    outstandingHeartbeats.set(0);
+                });
+                future.addListener(this::futureDone);
             } else {
                 IoWriteFuture future = session.writePacket(buf);
                 future.addListener(this::futureDone);
             }
-            heartbeatCount.incrementAndGet();
             return true;
         } catch (IOException | RuntimeException | Error e) {
             session.exceptionCaught(e);